+++ Wizzy/src/Wizzy/events/MouseEvent.h
@@ -8,19 +8,19 @@ namespace Wizzy {
 	class WZ_API MouseMovedEvent
 		: public Event {
 	private:
-		float m_x, m_y;
+		double m_x, m_y;
 	public:
 
 		EVENT_CLASS_TYPE(mouse_move)
 		EVENT_CLASS_CATEGORY(category_input | category_mouse)
 
-		inline MouseMovedEvent(float x, float y) {
+		inline MouseMovedEvent(double x, double y) {
 			m_x = x;
 			m_y = y;
 		}
 
-		inline const float& GetX() const { return m_x; }
-		inline const float& GetY() const { return m_y; }
+		inline const double& GetX() const { return m_x; }
+		inline const double& GetY() const { return m_y; }
 
 		inline virtual string ToString() const override {
 			return GetName() + ": X'" + std::to_string(m_x) + "'" + ": Y'" + std::to_string(m_y) + "'";
@@ -70,19 +70,19 @@ namespace Wizzy {
 	class WZ_API MouseScrolledEvent
 		: public Event {
 	private:
-		float m_xOffset, m_yOffset;
+		double m_xOffset, m_yOffset;
 	public:
 
 		EVENT_CLASS_TYPE(mouse_scroll)
 		EVENT_CLASS_CATEGORY(category_input | category_mouse)
 
-		inline MouseScrolledEvent(float xOffset, float yOffset) {
+		inline MouseScrolledEvent(double xOffset, double yOffset) {
 			m_xOffset = xOffset;
 			m_yOffset = yOffset;
 		}
 
-		inline const float& GetXOffset() const { return m_xOffset; }
-		inline const float& GetYOffset() const { return m_yOffset; }
+		inline const double& GetXOffset() const { return m_xOffset; }
+		inline const double& GetYOffset() const { return m_yOffset; }
 
 		inline virtual string ToString() const override {
 			return GetName() + ": X'" + std::to_string(m_xOffset) + "'" + ": Y'" + std::to_string(m_yOffset) + "'";
